requirements.txt
pyrogram
python-dotenv
openai
aiohttp
pydub
tgcrypto
